Add server URL override
Guessing the appropriate server URL is a bit more complicated
than just using the Host/Origin header so add a parameter for
an override. Useful if you have non-trivial setups (which most
setups are)Add server URL override

Guessing the appropriate server URL is a bit more complicated
than just using the Host/Origin header so add a parameter for
an override. Useful if you have non-trivial setups (which most
setups are)